VibeCycleBackend/requirements.txt
alembic
fastapi
fastmcp
ollama
passlib[bcrypt]
psycopg2
pydantic
pyjwt
python-decouple
python-multipart
sqlmodel
uvicorn
uv